Move base constants into a "globalconstants" package
Base directory of the code was getting a bit cluttered, so seperate the constants in the base folder into their own package.
