+++ src/main/java/com/team1165/robot/OdysseusState.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) 2025 Team Paradise - FRC 1165 (https://github.com/TeamParadise)
+ *
+ * Use of this source code is governed by the MIT License, which can be found in the LICENSE file at
+ * the root directory of this project.
+ */
+
+package com.team1165.robot;
+
+import com.team1165.robot.util.statemachine.State;
+
+public enum OdysseusState implements State {}
