+++ src/main/java/com/team1165/robot/Robot.java
@@ -5,6 +5,8 @@
 
 package com.team1165.robot;
 
+import com.team1165.robot.globalconstants.BuildConstants;
+import com.team1165.robot.globalconstants.Constants;
 import edu.wpi.first.wpilibj.Threads;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
