+++ src/main/java/com/team1165/robot/RobotContainer.java
@@ -18,6 +18,7 @@ import com.ctre.phoenix6.swerve.SwerveRequest;
 import com.ctre.phoenix6.swerve.SwerveRequest.SwerveDriveBrake;
 import com.team1165.robot.commands.elevator.ElevatorPosition;
 import com.team1165.robot.commands.flywheels.FlywheelsPercenmt;
+import com.team1165.robot.globalconstants.Constants;
 import com.team1165.robot.subsystems.drive.Drive;
 import com.team1165.robot.subsystems.drive.constants.DriveConstants;
 import com.team1165.robot.subsystems.drive.constants.TunerConstants;
@@ -32,8 +33,8 @@ import com.team1165.robot.subsystems.flywheels.Flywheels;
 import com.team1165.robot.subsystems.flywheels.io.FlywheelsIO;
 import com.team1165.robot.subsystems.flywheels.io.FlywheelsIOSparkMax;
 import com.team1165.robot.subsystems.roller.funnel.Funnel;
+import com.team1165.robot.subsystems.roller.funnel.FunnelConstants;
 import com.team1165.robot.subsystems.roller.funnel.FunnelState;
-import com.team1165.robot.subsystems.roller.funnel.constants.FunnelConstants;
 import com.team1165.robot.subsystems.roller.io.RollerIO;
 import com.team1165.robot.subsystems.roller.io.RollerIOSim;
 import com.team1165.robot.subsystems.roller.io.RollerIOSpark;
